requirements.txt
jupyterlab
openpyxl
xmltodict
dicttoxml
Pillow
iplantuml
ipydrawio
ipydrawio-export
lark
tqdm
python-Levenshtein